Prevent NetworkManager from managing /etc/resolv.conf
In CentOS 8, every time NetworkManager gets restarted, the content of
the /etc/resolv.conf file will be regenerated. This causes us problems
because a freshly kickstarted node has the gateway IP address configured
in the /etc/resolv.conf and we need this information in order to fetch
packages from the internet.
There is the possibility to prevent NetworkManager from managing the
/etc/resolv.conf file, so that the previous configuration could stay in
place. We have anyway Puppet code that will override this information
later.