Toggle interface also when ip is assigned to bridge (#33)
We have few cases in which the ansible config assigns the IP not to the
bond interface but to its bridge instead. In such cases, the current
toggling check fails because there are no ip addresses assigned to the
bond interfaces.
The change proposed in this commit checks if we have some bond slaves
whose ip address is the same as the ip address assigned to the bridge of
the bond. Such check is performed only if the original check is
completely skipped.