Check if bridge is defined before using it in checks
We have recently refined the handler so that it would disconnect and
connect any interface having the same ip as a corresponding bond
interface. This works fine in the nodes that define bond interfaces, but
not in the nodes that don't.
This commit adds a simple check to ensure that the interface handled is
indeed a bridge before going forward with further checks, which take the
fact of the interface being a bridge for granted.