Check if bridge is defined before using it in checks (#35)
We have recently refined the handler so that it would disconnect and
connect any interface having the same ip as a corresponding bond
interface. This works fine in the nodes that define bond interfaces, but
not in the nodes that don't.
This commit adds a simple check to ensure that the interface handled
is a bond that has indeed a bridge defined before going forward with
further checks, which take this fact for granted.