Fix conditions for check on bridge interfaces
Currently the check on bridge interfaces is performed only if no actions
have been taken on the bond interfaces in the previous task. This is
implemented by checking that the results obtained from the previous task
all contain the 'skipped: true' flag. Nevertheless, if an action is
taken in the previous task instead, then the 'skipped: true' flag is
completely absent (there is no 'skipped: false'), which breaks the
check.
We can obtain the same restriction by checking only the bonds for which
a bridge is defined in their data. Compute nodes do not have such bridge
field, and thus will be automatically excluded.
The commit also avoid to reapply the configuration for interfaces such
as 'virbr0'.